I ain't had a woman in years

+++ packages/typeslayer/src-tauri/src/validate/utils.rs
@@ -4,7 +4,7 @@ pub const CPU_PROFILE_FILENAME: &str = "tsc.cpuprofile";
 
 /// TypeId corresponds to z.number().int().positive().or(z.literal(-1)) in TS.
 /// We keep it as i64 to be safe with larger ids.
-pub type TypeId = i64;
+pub type TypeId = usize;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Position {
